// BulletTimeStudio/LandingPage/basketAnnoTool/src/app/socket/socket.interface.ts
export enum SOCKET_EVENTS {
    EVENT_FRAME_AVAILABLE = 'frame_available',
    EVENT_REQUEST_FRAME = 'request_frame',
    EVENT_GOAL = 'goal',
    EVENT_GOAL_COMPLETE = 'goal_complete',
    EVENT_DETECT_REPLY = 'detect_reply',
    EVENT_DETECT_DETAIL = 'detect_detail'
    // EVENT_START_RECORDING = 'start_recording',
}